fix: add high order function for getThemeProperty to be used properly in tests
