+++ packages/botonic-react/src/webchat/webchat.jsx
@@ -22,12 +22,12 @@ import { WebchatMessageList } from './message-list'
 import { WebchatReplies } from './replies'
 import { WebviewContainer } from './webview'
 import { msgToBotonic } from '../msg-to-botonic'
-import { isDev, staticAsset, getProperty } from '../utils'
+import { isDev, staticAsset, _getThemeProperty } from '../utils'
 import Logo from '../assets/botonic_react_logo100x100.png'
 import EmojiPicker from 'emoji-picker-react'
 import LogoMenu from '../assets/menuButton.svg'
 import LogoEmoji from '../assets/emojiButton.svg'
-import { CUSTOM_WEBCHAT_PROPERTIES, MIME_WHITELIST } from '../constants'
+import { MIME_WHITELIST } from '../constants'
 
 const getAttachmentType = fileType => {
   return Object.entries(MIME_WHITELIST)
@@ -74,22 +74,7 @@ export const Webchat = forwardRef((props, ref) => {
   const [emojiPickerOpened, setEmojiPickerOpened] = useState(false)
   const [attachment, setAttachment] = useState({})
 
-  /*
-   * returns the value of a property defined in bot's theme based on CUSTOM_WEBCHAT_PROPERTIES dictionary.
-   * it gives preference to nested defined properties over plain properties.
-   * if property doesn't exist, returns the defaultValue
-   */
-  const getThemeProperty = (property, defaultValue = undefined) => {
-    for (let [k, v] of Object.entries(CUSTOM_WEBCHAT_PROPERTIES)) {
-      if (v == property) {
-        let nestedProperty = getProperty(theme, v)
-        let plainProperty = getProperty(theme, k)
-        if (nestedProperty !== undefined) return nestedProperty
-        if (plainProperty !== undefined) return plainProperty
-        return defaultValue
-      }
-    }
-  }
+  const getThemeProperty = _getThemeProperty(theme)
 
   const handleAttachment = event => {
     setAttachment({
